Edited CompanyInternSearch delete method

+++ src/main/java/com/fcinar/interntrackingsystem/controller/CompanyInternSearchController.java
@@ -2,6 +2,7 @@ package com.fcinar.interntrackingsystem.controller;
 
 import com.fcinar.interntrackingsystem.dto.CompanyInternSearchDto;
 import com.fcinar.interntrackingsystem.dto.request.CreateCompanyInternSearchRequest;
+import com.fcinar.interntrackingsystem.model.CompanyInternSearch;
 import com.fcinar.interntrackingsystem.service.CompanyInternSearchService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -34,8 +35,13 @@ public class CompanyInternSearchController {
     @DeleteMapping("/intern-search/{companyId}")
     public ResponseEntity<HttpStatus> deleteCompanyInternSearchByCompanyId(@PathVariable UUID companyId) {
         try {
-            companyInternSearchService.deleteCompanyInternSearchByCompanyId(companyId);
-            return new ResponseEntity<>(HttpStatus.OK);
+            CompanyInternSearchDto companyInternSearch = companyInternSearchService.
+                    getCompanyInternSearchByCompanyId(companyId);
+            if (companyInternSearch != null) {
+                companyInternSearchService.deleteCompanyInternSearchByCompanyId(companyId);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
